Extracting channel module
