+++ packages/testkit-backend/src/channel/index.js
@@ -1,0 +1,7 @@
+import Channel from "./abstract"
+import SocketChannel from "./socket"
+
+export default Channel
+export {
+  SocketChannel
+}
